+++ apps/pioneer-lib/src/app/components/Chat/quest.tsx
@@ -1,0 +1,22 @@
+/*
+    Quest: 0.0.1 Become a Pioneer
+
+    Description:
+    This quest will guide you through the process of becoming a Pioneer. You will learn how to connect your wallet, pair your wallet, and interact with the Pioneer platform.
+
+
+    * Connect Wallet
+
+    * buy a Pioneer NFT
+      * ETH auction
+      * BASE auction
+      *open sea
+
+
+    Quizes:
+      is NFT
+
+ */
+
+
+