+++ packages/swapkit/api/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @swapkit/api
 
+## 3.7.59
+
+### Patch Changes
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/types@4.7.59
+
+## 3.7.58
+
+### Patch Changes
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/types@4.7.58
+
+## 3.7.57
+
+### Patch Changes
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/types@4.7.57
+
 ## 3.7.56
 
 ### Patch Changes