+++ packages/swapkit/core/CHANGELOG.md
@@ -1,5 +1,44 @@
 # @coinmasters/core
 
+## 11.0.59
+
+### Patch Changes
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/toolbox-cosmos@11.0.59
+  - @coinmasters/toolbox-utxo@11.0.59
+  - @coinmasters/types@4.7.59
+  - @coinmasters/toolbox-evm@11.0.59
+  - @coinmasters/api@3.7.59
+
+## 11.0.58
+
+### Patch Changes
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/toolbox-cosmos@11.0.58
+  - @coinmasters/toolbox-utxo@11.0.58
+  - @coinmasters/types@4.7.58
+  - @coinmasters/toolbox-evm@11.0.58
+  - @coinmasters/api@3.7.58
+
+## 11.0.57
+
+### Patch Changes
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/toolbox-cosmos@11.0.57
+  - @coinmasters/toolbox-utxo@11.0.57
+  - @coinmasters/types@4.7.57
+  - @coinmasters/toolbox-evm@11.0.57
+  - @coinmasters/api@3.7.57
+
 ## 11.0.56
 
 ### Patch Changes